+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Change max body size in nginx example config to match app limit.
+- Don't create invoice if recipient can't accept subscription payments.
 
 ## [1.12.0] - 2023-01-26